separated Cell2DataStructures to Java-backend-specific version and non-backend-specific version
